AG-8932 - Move enterprise typings into community.
