+++ packages/ag-charts-community/src/chart/interaction/animationOptions.ts
@@ -1,0 +1,6 @@
+export interface AgAnimationOptions {
+    /** Set to true to enable the animation module. */
+    enabled?: boolean;
+    /** The total duration of the animation for each series on initial load and updates */
+    duration?: number;
+}
