+++ packages/ag-charts-community/src/chart/series/cartesian/heatmapOptions.ts
@@ -1,12 +1,8 @@
-import type {
-    AgBaseSeriesOptions,
-    AgSeriesListeners,
-    AgSeriesTooltip,
-    AgCartesianSeriesTooltipRendererParams,
-    AgTooltipRendererResult,
-    CssColor,
-    PixelSize,
-} from 'ag-charts-community';
+import type { AgSeriesListeners } from '../../options/eventOptions';
+import type { AgSeriesTooltip, AgTooltipRendererResult } from '../../options/tooltipOptions';
+import type { CssColor, PixelSize } from '../../options/types';
+import type { AgBaseSeriesOptions } from '../seriesOptions';
+import type { AgCartesianSeriesTooltipRendererParams } from './cartesianOptions';
 
 export interface AgHeatmapSeriesFormatterParams<DatumType> {
     readonly datum: DatumType;
