+++ packages/ag-charts-community/src/chart/series/cartesian/rangeBarOptions.ts
@@ -1,16 +1,9 @@
-import type {
-    AgBaseSeriesOptions,
-    AgSeriesListeners,
-    AgSeriesHighlightStyle,
-    AgSeriesTooltip,
-    AgCartesianSeriesTooltipRendererParams,
-    AgTooltipRendererResult,
-    AgCartesianSeriesLabelOptions,
-    CssColor,
-    PixelSize,
-    Opacity,
-    AgDropShadowOptions,
-} from 'ag-charts-community';
+import type { AgDropShadowOptions } from '../../options/chartOptions';
+import type { AgSeriesListeners } from '../../options/eventOptions';
+import type { AgSeriesTooltip, AgTooltipRendererResult } from '../../options/tooltipOptions';
+import type { CssColor, Opacity, PixelSize } from '../../options/types';
+import type { AgBaseSeriesOptions, AgSeriesHighlightStyle } from '../seriesOptions';
+import type { AgCartesianSeriesLabelOptions, AgCartesianSeriesTooltipRendererParams } from './cartesianOptions';
 
 export interface AgRangeBarSeriesFormatterParams<DatumType> {
     readonly datum: DatumType;
