+++ packages/ag-charts-community/src/chart/series/polar/radarLineOptions.ts
@@ -1,4 +1,4 @@
-import type { AgBaseRadarSeriesOptions } from '../radar/typings';
+import type { AgBaseRadarSeriesOptions } from './radarOptions';
 
 export interface AgRadarLineSeriesOptions<DatumType = any> extends AgBaseRadarSeriesOptions<DatumType> {
     type?: 'radar-line';
