+++ packages/ag-charts-enterprise/src/animation/animationModule.ts
@@ -8,10 +8,3 @@ export const AnimationModule: _ModuleSupport.Module = {
     chartTypes: ['cartesian', 'polar'],
     instanceConstructor: Animation,
 };
-
-export interface AgAnimationOptions {
-    /** Set to true to enable the animation module. */
-    enabled?: boolean;
-    /** The total duration of the animation for each series on initial load and updates */
-    duration?: number;
-}
