+++ packages/ag-charts-enterprise/src/context-menu/contextMenuModule.ts
@@ -35,18 +35,3 @@ export function _disableAction(actionId: string) {
 export function _enableAction(actionId: string) {
     ContextMenu.enableAction(actionId);
 }
-
-export interface AgContextMenuOptions {
-    enabled?: boolean;
-    extraActions?: Array<AgContextMenuAction>;
-}
-
-export type AgContextMenuAction = {
-    label: string;
-    action: (params: AgContextMenuActionParams) => void;
-};
-
-export type AgContextMenuActionParams = {
-    datum?: any;
-    event: MouseEvent;
-};
