+++ packages/ag-charts-enterprise/src/heatmap/main.ts
@@ -1,2 +1,1 @@
 export { HeatmapModule } from './heatmapModule';
-export * from './typings';
