+++ packages/ag-charts-enterprise/src/polar-axes/angle-category/main.ts
@@ -1,2 +1,1 @@
 export { AngleCategoryAxisModule } from './angleCategoryAxisModule';
-export { AgAngleCategoryAxisOptions } from './typings';
