+++ packages/ag-charts-enterprise/src/polar-axes/radius-number/main.ts
@@ -1,2 +1,1 @@
 export { RadiusNumberAxisModule } from './radiusNumberAxisModule';
-export { AgRadiusNumberAxisOptions } from './typings';
