+++ packages/ag-charts-enterprise/src/polar-series/nightingale/main.ts
@@ -1,2 +1,1 @@
 export { NightingaleModule } from './nightingaleModule';
-export * from './typings';
