+++ packages/ag-charts-enterprise/src/polar-series/radar-area/main.ts
@@ -1,2 +1,1 @@
 export { RadarAreaModule } from './radarAreaModule';
-export * from './typings';
