+++ packages/ag-charts-enterprise/src/polar-series/radar-line/main.ts
@@ -1,2 +1,1 @@
 export { RadarLineModule } from './radarLineModule';
-export * from './typings';
