+++ packages/ag-charts-enterprise/src/polar-series/radial-column/main.ts
@@ -1,2 +1,1 @@
 export { RadialColumnModule } from './radialColumnModule';
-export * from './typings';
