+++ packages/ag-charts-enterprise/src/rangeBar/main.ts
@@ -1,2 +1,1 @@
 export { RangeBarModule, RangeColumnModule } from './rangeBarModule';
-export * from './typings';
