+++ packages/ag-charts-enterprise/src/rangeBar/rangeBar.ts
@@ -1,11 +1,12 @@
-import type { AgCartesianSeriesLabelFormatterParams, AgTooltipRendererResult } from 'ag-charts-community';
-import { _ModuleSupport, _Scale, _Scene, _Util } from 'ag-charts-community';
 import type {
+    AgCartesianSeriesLabelFormatterParams,
     AgRangeBarSeriesFormat,
     AgRangeBarSeriesFormatterParams,
-    AgRangeBarSeriesTooltipRendererParams,
     AgRangeBarSeriesLabelPlacement,
-} from './typings';
+    AgRangeBarSeriesTooltipRendererParams,
+    AgTooltipRendererResult,
+} from 'ag-charts-community';
+import { _ModuleSupport, _Scale, _Scene, _Util } from 'ag-charts-community';
 
 const {
     Validate,
