+++ packages/ag-charts-enterprise/src/waterfall/main.ts
@@ -1,2 +1,1 @@
 export { WaterfallBarModule, WaterfallColumnModule } from './waterfallModule';
-export * from './typings';
